Move session code to source root
