+++ src/__tests__/fake-resource.ts
@@ -1,5 +1,5 @@
 import Base, {ParamSet, ResourcePath} from '../base-rest-resource';
-import {SessionInterface} from '../auth/session/types';
+import {SessionInterface} from '../session/types';
 import {ApiVersion} from '../base-types';
 
 interface FakeResourceFindArgs {
