+++ src/__tests__/jest_projects/session_storage.jest.config.mjs
@@ -3,5 +3,5 @@ import baseConfig from './base.jest.config.mjs';
 export default {
   ...baseConfig,
   displayName: 'session_storage',
-  rootDir: '../../auth/session/storage',
+  rootDir: '../../session/storage',
 };
