+++ src/session/session.ts
@@ -1,5 +1,5 @@
-import {OnlineAccessInfo} from '../oauth/types';
-import {AuthScopes} from '../scopes';
+import {OnlineAccessInfo} from '../auth/oauth/types';
+import {AuthScopes} from '../auth/scopes';
 
 import {SessionInterface} from './types';
