+++ src/types.ts
@@ -1,2 +1,2 @@
 export * from './base-types';
-export * from './auth/session/types';
+export * from './session/types';
