+++ src/utils/__tests__/graphql_proxy.test.ts
@@ -3,7 +3,7 @@ import express, {Request, Response} from 'express';
 import request from 'supertest';
 
 import '../../__tests__/shopify-global';
-import {Session} from '../../auth/session';
+import {Session} from '../../session';
 import {InvalidSession, SessionNotFound} from '../../error';
 import graphqlProxy from '../graphql_proxy';
 import {config, setConfig} from '../../config';
