+++ src/utils/__tests__/load-offline-session.test.ts
@@ -1,7 +1,7 @@
 import '../../__tests__/shopify-global';
 import {config} from '../../config';
 import loadOfflineSession from '../load-offline-session';
-import {Session} from '../../auth/session/session';
+import {Session} from '../../session/session';
 import OAuth from '../../auth/oauth';
 
 describe('loadOfflineSession', () => {
